requirements.txt
# Core dependencies for the real-time data pipeline
pandas>=1.5.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Database connectors
snowflake-connector-python>=3.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Airflow
apache-airflow>=2.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0